Make todo & Edit cmds & Refact a bit
